moved CustomEquipmentModal to folder container
