+++ src/components/CustomEquipmentModal/CustomEquipmentModal.css
